requirements.txt
pyyaml
pymupdf
anthropic
openai
deepdiff
pdfplumber
pydantic
rapidfuzz
pandas
httpx
beautifulsoup4
playwright
tenacity
